engine_rocks: Rename properties_types to decode_properties
