+++ components/engine_rocks/src/decode_properties.rs
