+++ components/engine_rocks/src/lib.rs
@@ -72,8 +72,8 @@ pub use compat::*;
 mod compact_listener;
 pub use compact_listener::*;
 
-pub mod properties_types;
-pub use properties_types::*;
+pub mod decode_properties;
+pub use decode_properties::*;
 pub mod properties;
 pub use properties::*;
